+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- A list of all found credentials can be included in the report with degree
+  of detail >= 4
+
 ## [0.1.4] - 2022-06-27
 
 ### Added